Implement displaying of text/plain documents
This is done by detecting the content type as text/plain
and if it is then escaping all '<' in the document to
'&lt;' and inserting it in a document with a 'pre' element.

This is a prototype solution to check my view-source work
and to explore ways of doing text/plain support.